feat : modify NotProd for initial user setting

+++ src/main/java/com/example/MnM/boundedContext/member/service/MemberService.java
@@ -28,8 +28,8 @@ public class MemberService {
         String password = joinForm.getPassword();
         String email = joinForm.getEmail();
         String nickname = joinForm.getNickname();
-        Long height = joinForm.getHeight();
-        Long age = joinForm.getAge();
+        Integer height = joinForm.getHeight();
+        Integer age = joinForm.getAge();
         String locate =joinForm.getLocate();
         String gender = joinForm.getGender();
         String mbti = joinForm.getMbti();
